Moved memoization switch for type_node to be checked on access
The previous case had memoization check happen on import which means
NodeModelTestCase had to be initialized before NeoModel or it's subclasses
where imported (in order to set NEO4J_RUNNING_TESTS in settings).
This is not way tests are typically structured as imports should be at the
top not inside each test case.