Add a use_strings kwargs to ArrayProperty to store data as token separated string
use_strings kwarg when set to True will override default behavior of ArrayProperty
and store array data as a token separated string in neo4j instead of an array.
This is useful for users who interact with their neo4j instance with tools other
than neo4django, specifically REST as you cannot store empty arrays with REST.
The user can specify the token to use as the separator as well as the token to
use for escaping the separator with the values themselves. to_neo() will take care
of escaping the separator within the values themselves.